requirements.txt
# Core dependencies
python-telegram-bot[job-queue]==21.0.1
python-dotenv==1.0.1

# System monitoring
psutil==5.9.8

# Docker management
docker==7.0.0

# Data visualization
matplotlib==3.8.3
seaborn==0.13.2
Pillow==10.2.0

# Scheduling
APScheduler==3.10.4

# Async utilities
aiofiles==23.2.1

# Type checking
pydantic==2.6.1
pydantic-settings==2.1.0

# Development dependencies
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
ruff==0.2.1
mypy==1.8.0
